scsi: ufs: qcom: Properly size buffer for wq name
The intent of the code was probably to name a work queue
"ufs_mgc_hibern8_work_#" where # is replaced by the host_no.
However, due to the inadequate size of the string buffer, the name was
cut short to just "ufs_mgc_hibern8_work". Let's increase the buffer size
to provide enough space for the host_no.

Fixes: 9f79b4b ("scsi: ufs: add manual_gc in sysfs with UFS 3.1")
Change-Id: Icf99adeb8d4b24a748ab8043f61aa40ab6c0a9a3